PDG id now passed as long (for HI purposes), less hardcoded parameters in fortran kt process

+++ CepGen/Event/Particle.cpp
@@ -192,7 +192,7 @@ namespace cepgen
   }
 
   void
-  Particle::setPdgId( short pdg )
+  Particle::setPdgId( long pdg )
   {
     pdg_id_ = abs( pdg );
     try {
